# Work around a startup stall caused by expo-image on low-end Android