@gaearon 2023-11-03
Filed Expo bug: https://github.com/expo/expo/issues/25201